+++ controllers/PermissionController.php
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * @copyright Copyright &copy; Gogodigital Srls
+ * @company Gogodigital Srls - Wide ICT Solutions
+ * @website http://www.gogodigital.it
+ * @github https://github.com/cinghie/yii2-user-extended
+ * @license GNU GENERAL PUBLIC LICENSE VERSION 3
+ * @package yii2-user-extended
+ * @version 0.6.4
+ */
+
+namespace cinghie\userextended\controllers;
+
+use dektrium\rbac\controllers\PermissionController as BasePermissionController;
+use yii\filters\VerbFilter;
+
+/**
+ * Permission CRUD with POST-only delete (upstream ItemController has no VerbFilter).
+ */
+class PermissionController extends BasePermissionController
+{
+	/**
+	 * @inheritdoc
+	 */
+	public function behaviors()
+	{
+		return array_merge(parent::behaviors(), [
+			'verbs' => [
+				'class' => VerbFilter::class,
+				'actions' => [
+					'delete' => ['POST'],
+				],
+			],
+		]);
+	}
+}