+++ views/adminlte/security/login.php
@@ -42,6 +42,7 @@ $fieldOptions2 = [
 
         <?php $form = ActiveForm::begin([
             'id' => 'login-form',
+            'enableCsrfValidation'   => true,
             // AJAX validation burns single-use Turnstile tokens before the real submit
             'enableAjaxValidation'   => !\cinghie\userextended\helpers\TurnstileVerifier::shouldProtectLogin(),
             'enableClientValidation' => false,