+++ views/registration/_one_column.php
@@ -26,6 +26,7 @@ use yii\helpers\Html;
 
 	            <?php $form = ActiveForm::begin([
 		            'id' => 'registration-form',
+		            'enableCsrfValidation' => true,
 		            'enableAjaxValidation' => false,
 		            'enableClientValidation' => true,
 	            ]) ?>