+++ views/role/index.php
@@ -79,6 +79,16 @@ $this->params['breadcrumbs'][] = $this->title;
 				    'urlCreator' => function ($action, $model) {
 					    return Url::to(['/rbac/role/' . $action, 'name' => $model['name']]);
 				    },
+				    'buttons' => [
+					    'delete' => function ($url) {
+						    return Html::a('<span class="glyphicon glyphicon-trash"></span>', $url, [
+							    'title' => Yii::t('yii', 'Delete'),
+							    'data-confirm' => Yii::t('yii', 'Are you sure you want to delete this item?'),
+							    'data-method' => 'post',
+							    'data-pjax' => '0',
+						    ]);
+					    },
+				    ],
 			    ]
 		    ],
 		    'responsive' => true,